+++ data-plane/core/slim/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0](https://github.com/agntcy/slim/compare/slim-v0.4.0...slim-v0.5.0) - 2025-09-17
+
+### Fixed
+
+- use duration-string in place of duration-str ([#683](https://github.com/agntcy/slim/pull/683))
+
 ## [0.4.0](https://github.com/agntcy/slim/compare/slim-v0.3.15...slim-v0.4.0) - 2025-07-31
 
 ### Other