+++ data-plane/integrations/mcp-proxy/CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.7](https://github.com/agntcy/slim/compare/slim-mcp-proxy-v0.1.6...slim-mcp-proxy-v0.1.7) - 2025-09-18
+
+### Fixed
+
+- *(python-bindings)* default crypto provider initialization for Reqwest crate ([#706](https://github.com/agntcy/slim/pull/706))
+- use duration-string in place of duration-str ([#683](https://github.com/agntcy/slim/pull/683))
+
+### Other
+
+- release dataplane v0.5.0 ([#508](https://github.com/agntcy/slim/pull/508))
+- *(data-plane)* move to clang-19 ([#662](https://github.com/agntcy/slim/pull/662))
+- *(python-bindings)* improve folder structure and Taskfile division ([#628](https://github.com/agntcy/slim/pull/628))
+
 ## [0.1.6](https://github.com/agntcy/slim/compare/slim-mcp-proxy-v0.1.5...slim-mcp-proxy-v0.1.6) - 2025-07-31
 
 ### Added